Update all file paths to be relative.

+++ src/main/java/com/ibm/cldk/utils/BuildProject.java
@@ -46,7 +46,7 @@ public class BuildProject {
         String gradleSystemCommand = Arrays.stream(System.getenv("PATH").split(System.getProperty("path.separator"))).map(path -> new File(path, System.getProperty("os.name").toLowerCase().contains("windows") ? "gradle.bat" : "gradle")).filter(File::exists).findFirst().map(File::getAbsolutePath).orElse(null);
         File gradleWrapper = System.getProperty("os.name").toLowerCase().contains("windows") ? new File(projectRootPom, "gradlew.bat") : new File(projectRootPom, "gradlew");
 
-        return commandExists(gradleWrapper.getAbsoluteFile()).getKey() ? gradleWrapper.getAbsoluteFile().toString() : gradleSystemCommand;
+        return commandExists(gradleWrapper.getAbsoluteFile()).getKey() ? gradleWrapper.getAbsoluteFile()    .toString() : gradleSystemCommand;
     }
 
     public static Path tempInitScript;
@@ -159,13 +159,13 @@ public class BuildProject {
     }
 
     private static boolean buildProject(String projectPath, String build) {
-        File pomFile = new File(projectPath, "pom.xml");
+        File pomFile = new File(String.valueOf(Paths.get(projectPath).toAbsolutePath()), "pom.xml");
         if (build == null) {
             return true;
         } else if (build.equals("auto")) {
             if (pomFile.exists()) {
                 Log.info("Found pom.xml in the project directory. Using Maven to build the project.");
-                return mavenBuild(projectPath); // Use Maven if pom.xml exists
+                return mavenBuild(Paths.get(projectPath).toAbsolutePath().toString()); // Use Maven if pom.xml exists
             } else {
                 Log.info("Did not find a pom.xml in the project directory. Using Gradle to build the project.");
                 return gradleBuild(projectPath); // Otherwise, use Gradle
@@ -204,14 +204,14 @@ public class BuildProject {
      * Downloads library dependency jars of the given project so that the jars can be used
      * for type resolution during symbol table creation.
      *
-     * @param projectPath Path to the project under analysis
+     * @param projectPath Path to the project under javaee
      * @return true if dependency download succeeds; false otherwise
      */
     public static boolean downloadLibraryDependencies(String projectPath, String projectRootPom) throws IOException {
         // created download dir if it does not exist
         String projectRoot = projectRootPom != null ? projectRootPom : projectPath;
 
-        File pomFile = new File(projectRoot, "pom.xml");
+        File pomFile = new File((new File(projectRoot)).getAbsoluteFile(), "pom.xml");
         if (pomFile.exists()) {
             libDownloadPath = Paths.get(projectPath, "target", LIB_DEPS_DOWNLOAD_DIR).toAbsolutePath();
             if (mkLibDepDirs(projectPath))
@@ -231,21 +231,21 @@ public class BuildProject {
                         ));
             }
             Log.info("Found pom.xml in the project directory. Using Maven to download dependencies.");
-            String[] mavenCommand = {MAVEN_CMD, "--no-transfer-progress", "-f", Paths.get(projectRoot, "pom.xml").toString(), "dependency:copy-dependencies", "-DoutputDirectory=" + libDownloadPath.toString()};
+            String[] mavenCommand = {MAVEN_CMD, "--no-transfer-progress", "-f", Paths.get(projectRoot, "pom.xml").toAbsolutePath().toString(), "dependency:copy-dependencies", "-DoutputDirectory=" + libDownloadPath.toString()};
             return buildWithTool(mavenCommand);
         } else if (new File(projectRoot, "build.gradle").exists() || new File(projectRoot, "build.gradle.kts").exists()) {
-            if (GRADLE_CMD == null || !commandExists(new File(GRADLE_CMD)).getKey()) {
-                libDownloadPath = Paths.get(projectPath, "build", LIB_DEPS_DOWNLOAD_DIR).toAbsolutePath();
-                if (mkLibDepDirs(projectPath))
-                    Log.debug("Dependencies found/created in " + libDownloadPath);
-                else
-                    throw new IllegalStateException("Error creating library dependency directory in " + libDownloadPath);
+            libDownloadPath = Paths.get(projectPath, "build", LIB_DEPS_DOWNLOAD_DIR).toAbsolutePath();
+            if (mkLibDepDirs(projectPath))
+                Log.debug("Dependencies found/created in " + libDownloadPath);
+            else
+                throw new IllegalStateException("Error creating library dependency directory in " + libDownloadPath);
 
+            if (GRADLE_CMD == null || !commandExists(new File(GRADLE_CMD)).getKey()) {
                 String msg = GRADLE_CMD == null ?
                         "Could not find Gradle or valid Gradle Wrapper" :
                         MessageFormat.format("Could not verify that {0} exists", GRADLE_CMD);
                 Log.error(msg);
-                throw new IllegalStateException("Unable to execute Maven command. " +
+                throw new IllegalStateException("Unable to execute Gradle command. " +
                         (GRADLE_CMD == null ?
                                 "Could not find Gradle or valid Gradle Wrapper" :
                                 "Attempt failed with message\n" + commandExists(new File(GRADLE_CMD)).getValue()
@@ -271,8 +271,10 @@ public class BuildProject {
         if (libDownloadPath != null) {
             Log.info("Cleaning up library dependency directory: " + libDownloadPath);
             try {
-                Files.walk(libDownloadPath).filter(Files::isRegularFile).map(Path::toFile).forEach(File::delete);
-                Files.delete(libDownloadPath);
+                if (libDownloadPath.toFile().getAbsoluteFile().exists()) {
+                    Files.walk(libDownloadPath).filter(Files::isRegularFile).map(Path::toFile).forEach(File::delete);
+                    Files.delete(libDownloadPath);
+                }
             } catch (IOException e) {
                 Log.warn("Unable to fully delete library dependency directory: " + e.getMessage());
             }
